test and document blake3
